:ok_hand: Switch from welcome notification to modal
In this case the modal makes more sense as we want the user to actively confirm they are happy with the assigned default office
