+++ app/javascript/components/App/index.js
@@ -1,11 +1,16 @@
-import React, { Component } from 'react'
+import React from 'react'
 import getMuiTheme from 'material-ui/styles/getMuiTheme'
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider'
+import { ThemeProvider as GardenThemeProvider } from '@zendeskgarden/react-theming'
 
 import UserContext from '../../context/UserContext'
 import Loading from 'components/LoadingIcon'
 import Header from 'components/Header'
-import Notifications from 'components/Notifications'
+import WelcomeModal from 'components/WelcomeModal'
+
+import '@zendeskgarden/react-modals/dist/styles.css'
+import '@zendeskgarden/react-buttons/dist/styles.css'
+import '@zendeskgarden/react-tags/dist/styles.css'
 
 const MuiTheme = {
   fontFamily:
@@ -19,23 +24,25 @@ const MuiTheme = {
 
 const App = ({ loading, currentUser, offices, userPopover, toggleUserPopover, updateUserOffice, children }) => (
   <UserContext.Provider value={currentUser}>
-    <MuiThemeProvider muiTheme={getMuiTheme(MuiTheme)}>
-      {loading ? (
-        <Loading />
-      ) : (
-        <div>
-          <Header
-            currentUser={currentUser}
-            offices={offices}
-            togglePopover={e => toggleUserPopover(e.currentTarget)}
-            popover={userPopover}
-            handleOfficeSelect={office => updateUserOffice(currentUser, office)}
-          />
-          {currentUser && <Notifications />}
-          {children}
-        </div>
-      )}
-    </MuiThemeProvider>
+    <GardenThemeProvider>
+      <MuiThemeProvider muiTheme={getMuiTheme(MuiTheme)}>
+        {loading ? (
+          <Loading />
+        ) : (
+          <div>
+            <Header
+              currentUser={currentUser}
+              offices={offices}
+              togglePopover={e => toggleUserPopover(e.currentTarget)}
+              popover={userPopover}
+              handleOfficeSelect={office => updateUserOffice(currentUser, office)}
+            />
+            <WelcomeModal />
+            {children}
+          </div>
+        )}
+      </MuiThemeProvider>
+    </GardenThemeProvider>
   </UserContext.Provider>
 )
