+++ app/javascript/components/WelcomeModal/mutations/confirmProfileSettings.gql
