+++ package.json
@@ -65,10 +65,11 @@
   "dependencies": {
     "@apollo/react-hooks": "^3.1.3",
     "@babel/runtime-corejs2": "^7.6.3",
+    "@zendeskgarden/react-buttons": "^7.1.4",
     "@zendeskgarden/react-dropdowns": "^7.1.4",
     "@zendeskgarden/react-forms": "^7.1.4",
     "@zendeskgarden/react-grid": "^7.1.4",
-    "@zendeskgarden/react-notifications": "^7.1.4",
+    "@zendeskgarden/react-modals": "^7.1.4",
     "@zendeskgarden/react-tags": "^7.1.4",
     "@zendeskgarden/react-theming": "^7.1.4",
     "actioncable": "^5.1.4",
